fix/second-pr-review
-Created the "controllers" folder and moved the login and signup logic to the files present in it
