+++ backend/routes/login.js
@@ -1,37 +1,8 @@
 const express = require("express");
-const fs = require("fs");
-const path = require("path");
+const { loginUser } = require("../controllers/loginController");
 
 const router = express.Router();
-const dbFilePath = path.join(__dirname, "../db.json");
 
-const readDatabase = () => JSON.parse(fs.readFileSync(dbFilePath, "utf-8"));
-
-router.post("/login", (req, res) => {
-    try {
-        const { email, password } = req.body;
-
-        if (!email || !password) {
-            return res.status(400).json({ error: "Email and password are required" });
-        }
-
-        let database = readDatabase();
-
-        if (!database.customers) {
-            return res.status(400).json({ error: "No users found. Please sign up first." });
-        }
-
-        const user = database.customers.find((user) => user.email === email);
-        if (!user || user.password !== password) {
-            return res.status(401).json({ error: "Invalid email or password" });
-        }
-
-        res.status(200).json({ message: "Login successful", user });
-
-    } catch (error) {
-        console.error("Login Error:", error);
-        res.status(500).json({ error: "Internal server error" });
-    }
-});
+router.post("/login", loginUser);
 
 module.exports = router;
