+++ backend/routes/signup.js
@@ -1,38 +1,8 @@
 const express = require("express");
-const fs = require("fs");
-const path = require("path");
+const { signupUser } = require("../controllers/singupController");
 
 const router = express.Router();
-const dbFilePath = path.join(__dirname, "../db.json");
 
-const readDatabase = () => JSON.parse(fs.readFileSync(dbFilePath, "utf-8"));
-const writeDatabase = (data) => fs.writeFileSync(dbFilePath, JSON.stringify(data, null, 2));
-
-router.post("/signup", (req, res) => {
-    try {
-        const { name, email, phone, password } = req.body;
-
-        if (!name || !email || !phone || !password) {
-            return res.status(400).json({ error: "All fields are required" });
-        }
-
-        let database = readDatabase();
-        database.customers = database.customers || [];
-
-        if (database.customers.find((user) => user.email === email)) {
-            return res.status(400).json({ error: "Email already registered" });
-        }
-
-        const newUser = { id: database.customers.length + 1, name, email, phone, password };
-        database.customers.push(newUser);
-        writeDatabase(database);
-
-        res.status(201).json({ message: "Signup successful" });
-
-    } catch (error) {
-        console.error("Signup Error:", error);
-        res.status(500).json({ error: "Internal server error" });
-    }
-});
+router.post("/signup", signupUser);
 
 module.exports = router;
